requirements.txt
streamlit
crewai
python-dotenv
openai>=1.0.0 